Fix inconsistency with how wallet account is set
